This commit contains some pre-commit ironing up
and majorly roll the test/rmq/ back to the previous
organization in order to clearly being reviewed.

To make the value return from ProcessController be used
in coroutine, it should be convert to awaitable future.

ProcessController control the process by kiwipy through
RabbitMQ. The return value of controller method
is of the concurrent.futures.Future type which
if not awaitable. Therefore, I convert it into
asyncio.Future by `asyncio.wrap_future` and then await
it till the future get set. After this change,
plum can work with kiwipy develop branch and also
the previous version 0.5.3.
However, the API may be a little different. Different
from tornado's future, which can be yield when the
future is done.