Remote process controller testcase
 - test cases are reorganized
 - make sure every background runnig process is terminated before the
   test exit (communicator is closed)

Since `@pytest.mark.asyncio` not allowed to be
nested in the unittest.TestCase class, async_controller refer to
the  `RemoteProcessController` while sync_controller refer
to `RemoteProcessThreadController`.

For tests which have process launched in the background the process
should be terminated before the test exit. If not the case, when the
test exit the communicator will be closed, but the background running
process is still running, which will cause an unexpected exception
(which will not fail the test but printed to the stdout). Therefore, for
tests `test_play` and `test_status`, kill the running process at the end
of the test.