ensure task run in create_task is a coroutine
When the `LoopCommunicator` used, callback registered as subscriber and
converted to be scheduled in the given event loop, the converted
callback is run through `create_task` in `futures.py`. `create_task`
receive a task and schedule it in the event loop. If the function pass
to `create_task` is not a coroutine function one need wrap it to a coroutine by `utils.ensure_coroutine`.
Following function type are possible: 1) class with method `__call__`
defined, for example the `ProcessLauncher` can the subscriber for
processes controlling. 2) the `functool.partial` function which not
recognized as coroutine function in python <3.8, so need check `fn.func`
instead.

also this commit include:

- back compatible with kiwipy<0.6.0
- iron up the code