capture and return `proc.future.result()` exception
In coroutine `_contine` and `_launch` the result of the process is
returned as `proc.future().result()`. However, when process killed or
terminated with exception, the `proc.future()` will be set to exception
and `proc.future().result()` will raise exception. The temporary design
is to return the exception rather than raise it and propagate back to
the caller.